Improvements regarding null values
`EntityEqualityComparer` now handles null arguments gracefully
Added an additional null check to `ModerationService.GuildMemberUpdatedHandler`

+++ src/Shinoa/Services/TimedServices/ModerationService.cs
@@ -259,17 +259,19 @@ namespace Shinoa.Services.TimedServices
             using (var db = new ModerationContext(dbOptions))
             {
                 var guildBinding = await db.GuildBindings.FindAsync(userOld.Guild.Id.ToString());
-                var comparer = new EntityEqualityComparer();
+                if (guildBinding == null) return;
                 
+                var comparer = new EntityEqualityComparer();
                 var muteRoleRemoved = userOld.Roles.Except(userNew.Roles, comparer).Contains(guildBinding.Role, comparer);
+                var muteRoleAdded = userNew.Roles.Except(userOld.Roles, comparer).Contains(guildBinding.Role, comparer);
                 var binding = await db.GuildUserMuteBindings.FindAsync(userOld.Guild.Id.ToString(), userOld.Id.ToString());
+
                 if (muteRoleRemoved && binding != null)
                 {
                     db.Remove(binding);
                 }
 
-                var muteRoleAdded = userNew.Roles.Except(userOld.Roles, comparer).Contains(guildBinding.Role, comparer);
-                if (muteRoleAdded && binding == null) // Check if binding exists, otherwise double create
+                if (muteRoleAdded && binding == null) // Check if binding exists, to prevent double create
                 {
                     binding = new GuildUserMuteBinding
                     {
